Refactor SocketHandler frameworks
Description: This commit includes changes to the SocketHandler frameworks. The BlueBoat.js file has been modified to indicate that it is taken from the game code but will not be converted to TypeScript. Additionally, the export.ts file has been updated to include code mostly taken from the game code and converted to TypeScript.